+++ test/unit/primary.test.ts
@@ -234,7 +234,7 @@ config.parallel('primary.test.js', () => {
                     const obj = schemaObjects.simpleHuman();
                     await c.insert(obj);
                     const doc = await c.findOne().exec(true);
-                    await doc.atomicSet('firstName', 'foobar');
+                    await doc.atomicPatch({ firstName: 'foobar' });
                     const doc2 = await c.findOne().exec(true);
 
                     assert.strictEqual(doc2.get('firstName'), 'foobar');
@@ -253,7 +253,7 @@ config.parallel('primary.test.js', () => {
                     const doc = await c.findOne().exec(true);
                     let value;
                     const sub = doc.get$('firstName').subscribe((newVal: any) => value = newVal);
-                    await doc.atomicSet('firstName', 'foobar');
+                    await doc.atomicPatch({ firstName: 'foobar' });
                     await promiseWait(10);
                     assert.strictEqual(value, 'foobar');
                     sub.unsubscribe();
@@ -300,7 +300,7 @@ config.parallel('primary.test.js', () => {
                         if (count >= 2) pW8.resolve();
                     });
                     const doc2 = await c2.findOne().exec(true);
-                    await doc2.atomicSet('firstName', 'foobar');
+                    await doc2.atomicPatch({ firstName: 'foobar' });
                     await pW8.promise;
                     await AsyncTestUtil.waitUntil(() => value === 'foobar');
                     assert.strictEqual(count, 2);